refactor(ui): split selectors into files
